General optimizations (warnings eliminated)

+++ Areas/BDM_ConfigurationArea.h
@@ -90,7 +90,7 @@ public:
 	void readDeviceFile(const char file_name[], std::ostream &local_file);
 	void writeDeviceFile(const char file_name[], std::istream& data);
 
-	const size_t	m_large_buf		= 65536;
+	const uint32_t	m_large_buf		= 65536;
 	const uint32_t	m_cbWriteMax	= 1024;
 	const uint32_t	m_cbReadMax		= 1024;
